requirement.txt
langchain
langchain-openai
langchain-community
openai
qdrant-client
python-dotenv
fastapi
uvicorn
sqlalchemy
pydantic
dateparser